backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
joblib==1.5.3
numpy==2.4.3
pandas==3.0.1
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
scikit-learn==1.5.1
scipy==1.17.1
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
